pyproject.toml
[project]
name = "microfinance-env"
version = "0.2.0"
description = "Two-phase microfinance loan lifecycle RL environment built on OpenEnv"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    # OpenEnv core — provides Environment base class, create_fastapi_app,
    # EnvClient, HTTPEnvClient, and the CLI tooling
    "openenv-core>=0.3.0",

    # Web server stack (used by create_fastapi_app internally)
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",

    # Serialisation — Pydantic is the OpenEnv standard for models
    "pydantic>=2.0.0",

    # Async HTTP client used by EnvClient websocket transport
    "httpx>=0.27.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",   # for TestClient in tests
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

# Tell uv which package is the "project" so `uv sync --no-install-project`
# installs only deps (faster Docker layer caching)
[tool.uv]
package = true